Add a PoseView
