--- RoboKit Demo/Views/Socket View/Subviews/PoseView.swift
@@ -1,11 +1,0 @@
-//
-//  PoseView.swift
-//  RoboKit Demo
-//
-//  Created by Mariia Chemerys on 20.05.2025.
-//
-
-import SwiftUI
-import RoboKit
-
-
