+++ RoboKit Demo/Views/Socket View/Subviews/PoseView/RotationView.swift
@@ -1,0 +1,39 @@
+//
+//  RotationView.swift
+//  RoboKit Demo
+//
+//  Created by Mariia Chemerys on 22.05.2025.
+//
+
+import SwiftUI
+import RoboKit
+
+public struct RotationView: View {
+    @Environment(TCPClient.self) private var client: TCPClient
+    @Environment(InputSphereManager.self) private var inputSphereManager: InputSphereManager
+    @Environment(FormManager.self) private var formManager: FormManager
+    
+    public var body: some View {
+        
+        VStack(alignment: .leading, spacing: -10) {
+            
+            switch client.selectedDataMode {
+            case .live:
+                VStack() {
+                    ForEach(EulerAngle.allCases, id: \.self) { eulerAngle in
+                        RoboKit.InputSphereRotationSlider(eulerAngle: eulerAngle)
+                    }
+                    .frame(width: 380)
+                }
+                .padding(.top)
+            case .set:
+                VStack(spacing: -10) {
+                    ForEach(EulerAngle.allCases, id: \.self) { eulerAngle in
+                        RoboKit.FormRotationTextField(eulerAngle: eulerAngle)
+                    }
+                    .frame(width: 280)
+                }
+            }
+        }
+    }
+}
